vlm/requirement.txt
torch
transformers
scikit-learn
numpy
pandas
Pillow
tqdm
pyarrow
scipy